move adapters folder
