+++ test-resources/src/adapters/dynamoDbAdapter.ts
